fix: dto 수정.
